# Conflicts:
#	esp-src/esp-be/src/main/java/edu/esp/database/daos/AdminDAO.java
#	esp-src/esp-be/src/main/java/edu/esp/database/daos/InstructorDAO.java
#	esp-src/esp-be/src/main/java/edu/esp/database/daos/StudentDAO.java

+++ esp-src/esp-be/src/main/java/edu/esp/database/DBFacadeImp.java
@@ -42,7 +42,7 @@ public class DBFacadeImp {
         return this.studentDAO.readStudentById(studentId);
     }
 
-    public Admin loadAdmin(int adminId) {
+    public Admin loadAdmin(byte adminId) {
         return this.adminDAO.readAdminById(adminId);
     }
 
@@ -58,15 +58,15 @@ public class DBFacadeImp {
         return this.instructorDAO.deleteInstructorById(instructorId);
     }
 
-    public boolean deleteAdmin(int adminId){
+    public boolean deleteAdmin(byte adminId){
         return this.adminDAO.deleteAdminById(adminId);
     }
 
-    public boolean signUpInstructor(byte id, int hash, Instructor instructor) {
+    public boolean signUpInstructor(int id, int hash, Instructor instructor) {
         return this.instructorDAO.signUpInstructor(id, hash, instructor);
     }
 
-    public boolean signUpStudent(byte id, int hash, Student student) {
+    public boolean signUpStudent(int id, int hash, Student student) {
         return this.studentDAO.signUpStudent(id, hash, student);
     }
 }
