+++ esp-src/esp-be/src/main/java/edu/esp/database/daos/AdminDAO.java
@@ -2,11 +2,10 @@ package edu.esp.database.daos;
 
 import edu.esp.system_entities.system_users.Admin;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
-
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-
+import org.springframework.jdbc.core.RowMapper;
 import java.util.List;
 
 public class AdminDAO {
@@ -16,7 +15,7 @@ public class AdminDAO {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Admin readAdminById(int id) {
+    public Admin readAdminById(byte id) {
         try {
             String sql = """
                 SELECT *
@@ -63,7 +62,7 @@ public class AdminDAO {
 
     }
 
-    public boolean deleteAdminById(int id) {
+    public boolean deleteAdminById(byte id) {
         try{
             String sql = """
                     DELETE FROM sys_admin
